LayoutTests/imported/w3c/web-platform-tests/content-security-policy/worker-src/shared-self.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.


FAIL Same-origin dedicated worker allowed by 'self'. Can't find variable: SharedWorker
